fix(zkvms/zkm): Update SDK and remove hack
Recently ZKM introduced a bug, where SNARK_SETUP would crash, but the
witness would be generated. So we introduced a hack where we would run
SNARK_SETUP, ignore the crash and rerun the program again, without
SNARK_SETUP.

Now they fixed it, so with SNARK_SETUP, the witnesses would be generated
and the proving process will be done directly after.