fix(zkvmLib): Don't remove final entry in lockfile
The AWK script goes to the next line when a property is found. If the
lock file ends on a property, the next line will be "fetched", but that
is after the end of the file, so we'll never reach the printing logic.
We're adding a blank line to the end of the lock file, so we never hit
this edge case.