fix(zkvms_host_io): Truncate output file in emit_benchmark_results when appropriate
If we're not appending to a file with contents, then when we write our
data, we're going to be overwriting the old data. If the old data is
longer than the new one, we'll have some leftover text in the end, which
will cause a problem with JSON parsing.

We'll be truncating the file (i.e. emptying it out), but only when not
appending to it.